code/requirements.txt
streamlit
google-generativeai
google-adk
requests
